requirements.txt
requests
six
python-dotenv
beautifulsoup4
lxml